:recycle: Rename to detectLabel
