+++ packages/functions/src/index.ts
@@ -3,4 +3,4 @@ admin.initializeApp();
 
 exports.writeSpotImageURLOnResized = require("./functions/writeSpotImageURLOnResized/function");
 exports.setUserID = require("./functions/setUserID/function");
-exports.emotionalise = require("./functions/emotionalise/function");
+exports.detectLabel = require("./functions/detectLabel/function");
